feat: Add permission key to recreate new instance on permission change
