+++ PermissionManager/src/main/kotlin/com/infomaniak/core/permissionmanager/SupportedApiPermissionManagerState.kt
@@ -19,15 +19,12 @@ package com.infomaniak.core.permissionmanager
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.Stable
-import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.snapshotFlow
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.PermissionState
 import com.google.accompanist.permissions.isGranted
-import com.google.accompanist.permissions.shouldShowRationale
 import kotlinx.coroutines.CompletableJob
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.first
@@ -36,27 +33,11 @@ import kotlinx.coroutines.flow.first
 @OptIn(ExperimentalPermissionsApi::class)
 internal class SupportedApiPermissionManagerState(
     private val permissionState: PermissionState,
-    private val _shouldShowRationale: MutableState<Boolean>,
 ) : PermissionManagerState {
-    override val shouldShowRationale: Boolean by _shouldShowRationale
-
-    override fun askPermission() {
-        if (permissionState.status.isGranted) return
-
-        if (permissionState.status.shouldShowRationale) {
-            _shouldShowRationale.value = true
-        } else {
-            permissionState.launchPermissionRequest()
-        }
-    }
-
-    override fun dismissAndAskPermission() {
-        _shouldShowRationale.value = false
-        permissionState.launchPermissionRequest()
-    }
+    override fun askPermissionIfNeeded() = permissionState.launchPermissionRequest()
 
     @Composable
-    override fun waitUntilGranted(action: () -> Unit): () -> Unit = with(permissionState) {
+    override fun waitUntilPermissionGranted(action: () -> Unit): () -> Unit = with(permissionState) {
         val isGrantedFlow = remember(permission) { snapshotFlow { status.isGranted } }
         val actionFired: CompletableJob = remember(permission) { Job() }
